% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_doc__objects.R
\name{create_objects}
\alias{create_objects}
\alias{make_doc_position}
\alias{make_doc_range}
\title{Create objects for rstudioapi.

Create \code{"document_position"} and \code{"document_range"} objects.}
\usage{
make_doc_position(row, column)

make_doc_range(start.row, start.column, end.row, end.column)
}
\arguments{
\item{row, column, start.row, start.column, end.row, end.column}{(integer)\cr
Single positive number that indicates certain position.}
}
\value{
An object of appropriate class.
}
\description{
Create objects for rstudioapi.

Create \code{"document_position"} and \code{"document_range"} objects.
}
\examples{
make_doc_position(2, 5)

make_doc_range(2, 5, 2, 6)
}
