% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_blank_line.R
\name{ensure_blank_line}
\alias{ensure_blank_line}
\title{Add blank lines if needed.}
\usage{
ensure_blank_line(text, context = rs_get_context(), above = FALSE,
  below_first_row = FALSE, below_selection = FALSE)
}
\arguments{
\item{text}{(character) Text to modify.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}

\item{above}{(logical) If \code{TRUE}, checks for the blank line above the
selection and if it is missing, adds blank line before the \code{text}.}

\item{below_first_row}{(logical) If \code{TRUE}, checks for the blank line below
the first selected row and if it is missing, adds blank line after
the \code{text}.}

\item{below_selection}{(logical) If \code{TRUE}, checks for the blank line
below the last selected row and if it is missing, adds blank line after
the \code{text}.}
}
\value{
String.
}
\description{
Add blank lines if needed.
}
