% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_blank_line_needed_above.R
\name{is_blank_line_needed_above}
\alias{is_blank_line_needed_above}
\title{Check if a blank line is needed above.}
\usage{
is_blank_line_needed_above(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\value{
Logical value.
}
\description{
Check if a blank line should be added above the selection.
}
