% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_symbol.R
\name{make_spaces}
\alias{make_spaces}
\title{Generate spaces of certain length}
\usage{
make_spaces(n)
}
\arguments{
\item{n}{(integer) A vertor of integers indicating number of scoaaces to generate.}
}
\value{
A character vector of length \code{length(n)} with spaces.
}
\description{
Generate spaces of certain length
}
\examples{
make_spaces(c(2, NA))
}
