% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data--rs_document_context.R
\docType{data}
\name{rs_document_context}
\alias{rs_document_context}
\title{Data: RStudio decment context data.}
\description{
Data: RStudio decment context data.
}
\examples{
rs_document_context
}
\keyword{data}
