% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_enclose_selection_with.R
\name{rs_enclose_selection_with}
\alias{rs_enclose_selection_with}
\title{Enclose selection of text with indicated symbols.}
\usage{
rs_enclose_selection_with(
  symbol = "",
  symbol_before = symbol,
  symbol_after = symbol,
  trim = FALSE,
  context = rs_get_context()
)
}
\arguments{
\item{symbol}{(character) A sequence of symbols to add on both sides of selection.}

\item{symbol_before}{(character) A sequence of symbols to before the selection
(overrides value of \code{symbol}).}

\item{symbol_after}{(character) A sequence of symbols to add after the selection
(overrides value of \code{symbol}).}

\item{trim}{(logical) Flag if whitespace should be trimmed from both sides of the selection.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Enclose selection of text with certain symbols which usually have special
meaning in R Markdown files.
}
\details{
If text is selected, the cursor is placed after the modified selection.
If no text is selected, the cursor is placed between the inserted symbols.
}
