% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_enclose_selected_rows.R
\name{rs_enclose_with_lines}
\alias{rs_enclose_with_lines}
\alias{rs_enclose_selected_rows_with}
\alias{rs_enclose_first_row_with}
\title{Enclose setected rows with text.}
\usage{
rs_enclose_selected_rows_with(text_above = NA, text_below = NA,
  ensure_blank_above = FALSE, context = rs_get_context())

rs_enclose_first_row_with(text_above = NULL, text_below = NULL,
  ensure_blank_above = FALSE, context = rs_get_context())
}
\arguments{
\item{text_above}{(character) Text to be inserted above the selection.}

\item{text_below}{(character) Text to be inserted below the selection
(the first row).}

\item{ensure_blank_above}{(logical)
If \code{TRUE}, checks if there is a
blank line above the selection. If line is not blank, a blank line
is added. If \code{FALSE}, blank line is not added.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
Enclose selected rows with text: \itemize{
    \item \code{rs_enclose_all_with_lines()} - above and below all selected rows;
    \item \code{rs_enclose_first_row_with_lines()} - above and below the first selected row only.
}
}
