% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_index.R
\name{rs_get_index}
\alias{rs_get_index}
\alias{rs_get_index_first_selected_col}
\alias{rs_get_index_last_selected_col}
\alias{rs_get_index_first_selected_row}
\alias{rs_get_index_last_selected_row}
\alias{rs_get_index_selected_rows}
\title{Get indices related to the selection(s).}
\usage{
rs_get_index_first_selected_col(
  selection = c("all", "first", "last"),
  context = rs_get_context()
)

rs_get_index_last_selected_col(
  selection = c("all", "last", "first"),
  context = rs_get_context()
)

rs_get_index_first_selected_row(
  selection = c("first", "last", "all"),
  context = rs_get_context()
)

rs_get_index_last_selected_row(
  selection = c("last", "first", "all"),
  context = rs_get_context()
)

rs_get_index_selected_rows(context = rs_get_context())
}
\arguments{
\item{selection}{(string) \cr
String that indicates, which selection should be extracted.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Get indices related to the selection(s).
}
