% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_position.R
\name{rs_get_position}
\alias{rs_get_position}
\alias{rs_get_position_selection_start}
\alias{rs_get_position_selection_end}
\alias{rs_get_index_selection_start}
\alias{rs_get_index_selection_end}
\title{Get position(s) related to the selection(s).}
\usage{
rs_get_position_selection_start(selection = c("all", "first", "last"),
  context = rs_get_context())

rs_get_position_selection_end(selection = c("all", "first", "last"),
  context = rs_get_context())

rs_get_index_selection_start(selection = c("first", "last"),
  context = rs_get_context())

rs_get_index_selection_end(selection = c("last", "first"),
  context = rs_get_context())
}
\arguments{
\item{selection}{(string) \cr
String that indicates, which selection should be extracted.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\value{
A list of \code{docdument_position} objects.
}
\description{
Get position(s) related to the selection(s).
}
