% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_row_lengths}
\alias{rs_get_row_lengths}
\title{Get lengths of selected rows.}
\usage{
rs_get_row_lengths(row, end_row = NULL, context = rs_get_context())
}
\arguments{
\item{row}{(numeric) \cr
Index of the first row of interest of a vector of row indices.}

\item{end_row}{(numeric | \code{NULL}) \cr
Index of the last row of interest or \code{NULL}.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\value{
An integer vector with number of characters in each selection.
}
\description{
Calculate number of characters in each selected row.
}
