% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_row_range.R
\name{rs_get_row_range}
\alias{rs_get_row_range}
\alias{rs_get_row_range_w_newline}
\alias{rs_get_first_selected_row_range}
\title{Get row range of the indicated line}
\usage{
rs_get_row_range(row, context = rs_get_context())

rs_get_row_range_w_newline(row)

rs_get_first_selected_row_range(include_newline = FALSE,
  context = rs_get_context())
}
\arguments{
\item{row}{(ineger) row index.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}

\item{include_newline}{(logical) indicates if the new line symbol should be included in the range}
}
\value{
An object of class "document_range"
}
\description{
`rs_get_row_range()` gets range of the indicated row from the beginning to the end of the line (excluding the new line symbol).\cr
`rs_get_row_range_w_newline()` gets range of the indicated row from the beggining of the line to the beginning to the next line.\cr
`rs_get_first_selected_row_range()` gets range of the first selected line.
}
