% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_range}
\alias{rs_get_selection_range}
\title{Get range of selection}
\usage{
rs_get_selection_range(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
"document_range" object.
}
\description{
Get the range of the first selection.
}
