% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_text}
\alias{rs_get_selection_text}
\title{Get selection text.}
\usage{
rs_get_selection_text(
  selection = c("all", "first", "last"),
  as_list = FALSE,
  context = rs_get_context()
)
}
\arguments{
\item{selection}{(string) \cr
String that indicates, which selection should be extracted.}

\item{as_list}{(logical)
Flag, if result should be a list, if \code{selection} is either
\code{"first"} or \code{"last"}.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\value{
A character vector.
}
\description{
Get the text in either the first selection or all selections.
}
