% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_text.R
\name{rs_get_text}
\alias{rs_get_text}
\alias{rs_get_selected_rows}
\title{Get indicated lines}
\usage{
rs_get_text(start, end = start, context = rs_get_context())

rs_get_selected_rows(context = rs_get_context())
}
\arguments{
\item{start}{Either \code{"document_position"} object of vector with the coordinates (row and column) of the begining of the selection.}

\item{end}{Either \code{"document_position"} object of vector with the coordinates (row and column) of the end of the selection.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
Character vector (extracted strings).
}
\description{
Get indicated lines
}
