% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_insert_text.R
\name{rs_insert_at_row_start}
\alias{rs_insert_at_row_start}
\title{Insert text at the beginning of the row(s).}
\usage{
rs_insert_at_row_start(rows, text = NULL, id = rs_get_context()$id)
}
\arguments{
\item{rows}{(integer) \cr
The index(es) of the row(s) in the document.}

\item{text}{(character)  \cr
The text to insert. This should either be length one (in which case,
this text is applied to each specified row); otherwise, it should be
the same length as the \code{rows} vector.}

\item{id}{The document id. When \code{NULL} or blank,
the mutation will apply to the currently open, or last
focused, RStudio document. Use the \code{id} returned
from \code{\link{getActiveDocumentContext}()} to ensure
that the operation is applied on the intended document.}
}
\description{
Insert text at the beginning of the row(s).
}
