% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_insert_before_first_selected_row.R
\name{rs_insert_before_first_selected_row}
\alias{rs_insert_before_first_selected_row}
\title{Insert text at the beginning of the first selected row.}
\usage{
rs_insert_before_first_selected_row(
  text = "",
  ensure_blank_above = FALSE,
  context = rs_get_context()
)
}
\arguments{
\item{text}{(character) The text to insert.}

\item{ensure_blank_above}{(logical)
If \code{TRUE}, inserts blank line above the selection if the line is not blank.
If \code{FALSE}, no blank line is added.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Insert text at the beginning of the first selected row.
}
