% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_insert_at_row_start.R
\name{rs_insert_text}
\alias{rs_insert_text}
\title{Insert text at the cursor position.}
\usage{
rs_insert_text(text = NULL, context = rs_get_context(),
  spaces = FALSE)
}
\arguments{
\item{text}{(character) The text to insert.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}

\item{spaces}{(logical) If \code{TRUE}, ensures that text is
surrounded by spaces.}
}
\description{
Insert text at the cursor position.
}
