% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_replace_in_selection.R
\name{rs_replace_in_selection}
\alias{rs_replace_in_selection}
\alias{rs_replace_selection}
\alias{select_correct_range}
\title{Replace text in selection.}
\usage{
rs_replace_in_selection(
  pattern,
  replacement,
  fixed = TRUE,
  keep_selected = TRUE,
  selection = c("all", "first", "last"),
  context = rs_get_context()
)

rs_replace_selection(
  replacement,
  keep_selected = TRUE,
  selection = c("all", "first", "last"),
  context = rs_get_context()
)

select_correct_range(old_text, new_text, old_range, id = NULL)
}
\arguments{
\item{pattern}{(character) \cr
A pattern of text to be replaced.}

\item{replacement}{(character) \cr
The replacement text.}

\item{fixed}{(logical) \cr
If \code{TRUE}, the pattern is a fixed expression.
If \code{FALSE}, the pattern is a regular expression.}

\item{keep_selected}{(logical) \cr
Flag indicating if the selection should be kept after add-in is applied.}

\item{selection}{(string) \cr
String that indicates, which selection should be extracted.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}

\item{old_text}{(character) Original text.}

\item{new_text}{(character) Text after correction.}

\item{old_range}{\code{document_range} object. See \link[rstudioapi:document_range]{rstudioapi::document_range}.}

\item{id}{Document ID. See \code{\link[rstudioapi:getActiveDocumentContext]{getActiveDocumentContext()}}.}
}
\description{
Replace text in selection.
}
\details{
\code{select_correct_range()} correctly adjust current selection.
}
\seealso{
\code{\link[base:gsub]{base::gsub()}}
}
