% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-package.R
\docType{package}
\name{addin.tools}
\alias{addin.tools}
\title{Helper functions to create RStudio add-ins.}
\description{
Package \code{addin.tools} contains various functions that help to construct
"RStudio" add-ins.

\code{License:} MIT   \cr
\code{URL:} \url{https://GegznaV.github.io/addin.tools} \cr
\code{Bug reports and suggestions:}
     \url{https://github.com/GegznaV/addin.tools/issues} \cr
\code{Author:} Vilmantas Gegzna
}
