% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_visual_editor.R
\name{is_visual_editor}
\alias{is_visual_editor}
\alias{is_rmd_visual_mode}
\title{Check if Visual Editor (VE) is used}
\usage{
is_visual_editor()

is_rmd_visual_mode()
}
\value{
Logical:
\itemize{
\item \code{TRUE} if an active document is in Visual Editor mode and
current cursor position or selection is outside a code chunk.
\item \code{FALSE} if an active document is:
a.  not in the Visual Editor mode;
b.  in Visual Editor mode but current cursor position or
selection is inside a code chunk.
}
}
\description{
Check if RStudio addins should treat the current selection of active document
as being in RStudio Visual Editor mode (RS VE) or not.
}
\details{
The document \link[rstudioapi:getActiveDocumentContext]{context} from RS VE
is different than regular RStudio document context.

\code{is_rmd_visual_mode()} was renamed to \code{is_visual_editor()}.
The alias is kept for backward compatibility but issues a deprecation warning.
}
