% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated--rs_get_ind.R
\name{rs_get_ind}
\alias{rs_get_ind}
\alias{rs_get_selection_start_index}
\alias{rs_get_selection_end_index}
\alias{rs_get_first_selected_col_index}
\alias{rs_get_last_selected_col_index}
\alias{rs_get_first_selected_row_index}
\alias{rs_get_last_selected_row_index}
\alias{rs_get_selected_row_indexes}
\title{Get indices from the selection (DEPRECATED).}
\usage{
rs_get_selection_start_index(selection = c("first", "last"),
  context = rs_get_context())

rs_get_selection_end_index(selection = c("last", "first"),
  context = rs_get_context())

rs_get_first_selected_col_index(selection = c("first", "last"),
  context = rs_get_context())

rs_get_last_selected_col_index(selection = c("last", "first"),
  context = rs_get_context())

rs_get_first_selected_row_index(selection = c("first", "last"),
  context = rs_get_context())

rs_get_last_selected_row_index(selection = c("last", "first"),
  context = rs_get_context())

rs_get_selected_row_indexes(context = rs_get_context())
}
\arguments{
\item{selection}{(string) \cr
String, that indicates, which selection should be extracted.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
NOTE: These functions are DEPRECATED!

Get index of the first/last row/column in the selection.
}
