% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_n_selections}
\alias{rs_get_n_selections}
\title{Get number of selections.}
\usage{
rs_get_n_selections(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
Number of selections.
}
\description{
Get number of selections.
}
