% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_row_lengths}
\alias{rs_get_row_lengths}
\title{Get lengths of selected rows.}
\usage{
rs_get_row_lengths(row, end_row = NULL, context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
An integer vector with number of characters in each selection.
}
\description{
Calculate number of characters in each selected row.
}
