% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_length}
\alias{rs_get_selection_length}
\title{Get length of selection.}
\usage{
rs_get_selection_length(selection = c("all", "first", "last"),
  context = rs_get_context())
}
\arguments{
\item{selection}{(string) \cr
String, that indicates, which selection should be extracted.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
An integer vector with number of characters in each selection.
}
\description{
Calculate number of characters in each selection.
}
