% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_length}
\alias{rs_get_selection_length}
\title{Length of selection}
\usage{
rs_get_selection_length(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
Integer with number of characters in the first selection.
}
\description{
Calculate number of characters in the first selection.
}
