% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_range}
\alias{rs_get_selection_range}
\title{Get range of selection}
\usage{
rs_get_selection_range(context = rs_get_context(), type = c("first",
  "all"))
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
Either a "document_range" object or a list of those objects.
}
\description{
Get the range of the first/each selection.
}
