% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_range}
\alias{rs_get_selection_range}
\title{Get range of selection.}
\usage{
rs_get_selection_range(
  selection = c("all", "first", "last"),
  as_list = FALSE,
  context = rs_get_context()
)
}
\arguments{
\item{selection}{(string) \cr
String that indicates, which selection should be extracted.}

\item{as_list}{(locical) \cr
Indicates if output sould be returned as a list.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\value{
Either a "document_range" object, if \code{selection} is "first" or
        "last", and \code{as_list = TRUE}, or a list of those objects otherwise.
}
\description{
Get the range of the first/each selection.
}
