% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_text}
\alias{rs_get_selection_text}
\title{Get selection text}
\usage{
rs_get_selection_text(context = rs_get_context(), type = c("first",
  "all"))
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
A character vector.
}
\description{
Get the text in either the first selection or all selections
}
