% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_selection.R
\name{rs_get_selection_text}
\alias{rs_get_selection_text}
\title{Get selection text}
\usage{
rs_get_selection_text(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\value{
String.
}
\description{
Get the text in the first selection.
}
