% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_text.R
\name{rs_get_text}
\alias{rs_get_text}
\alias{rs_get_selected_rows}
\alias{rs_get_first_selected_row}
\alias{rs_get_first_selected_row_length}
\alias{rs_get_last_selected_row}
\alias{rs_get_last_selected_row_length}
\title{Get text in indicated lines.}
\usage{
rs_get_text(row, end_row = NULL, context = rs_get_context())

rs_get_selected_rows(context = rs_get_context())

rs_get_first_selected_row(context = rs_get_context())

rs_get_first_selected_row_length(context = rs_get_context())

rs_get_last_selected_row(context = rs_get_context())

rs_get_last_selected_row_length(context = rs_get_context())
}
\arguments{
\item{row}{(integer) Index of the first row of interest.}

\item{end_row}{(integer) Index of the last row of interest.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\value{
Character vector with attribute \code{row_numbers} indicating which rows were returned.
}
\description{
Get text in all indicated consequtive lines.
}
