% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_ind.R, R/rs_get_text.R
\name{rs_get_text}
\alias{rs_get_text}
\alias{rs_get_selected_rows}
\alias{rs_get_text}
\title{Get indices from the selection.}
\usage{
rs_get_selected_rows(context = rs_get_context())

rs_get_text(start, end = start, context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}

\item{start}{Either \code{"document_position"} object of vector with the coordinates (row and column) of the begining of the selection.}

\item{end}{Either \code{"document_position"} object of vector with the coordinates (row and column) of the end of the selection.}
}
\value{
Extracted string(s).
}
\description{
Get index of the first/last row/column in the selection.

Get text
}
