% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_replace_in_selection.R
\name{rs_replace_in_selection}
\alias{rs_replace_in_selection}
\alias{rs_replace_selection}
\title{Replace text in selection.}
\usage{
rs_replace_in_selection(pattern, replacement, fixed = TRUE,
  keep_selected = TRUE, context = rs_get_context())

rs_replace_selection(replacement, keep_selected = TRUE,
  context = rs_get_context())
}
\arguments{
\item{pattern}{(character) A pattern of text to be replaced.}

\item{replacement}{(character) The replacement text.}

\item{fixed}{(logical)
If \code{TRUE}, the pattern is a fixed expression.
If \code{FALSE}, the pattern is a regular expression.}

\item{keep_selected}{(logical) Flag indicating if the selection should be
kept after add-in was applied.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
Replace text in selection.
}
\seealso{
\code{\link[base]{gsub}}
}
