% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_select_rows.R
\name{rs_select_rows}
\alias{rs_select_rows}
\title{Select rows}
\usage{
rs_select_rows(first, last, context = rs_get_context())
}
\arguments{
\item{first}{(integer) Index of the first row to select.}

\item{last}{(integer) Index of the last row to select.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
Select rows
}
