% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_select_rows.R
\name{rs_select_rows}
\alias{rs_select_rows}
\alias{rs_select_all_selected_rows}
\alias{rs_select_first_selected_row}
\alias{rs_select_last_selected_row}
\alias{rs_deselect_range}
\title{Select rows}
\usage{
rs_select_rows(first, last = NULL, context = rs_get_context())

rs_select_all_selected_rows(context = rs_get_context())

rs_select_first_selected_row(context = rs_get_context())

rs_select_last_selected_row(context = rs_get_context())

rs_deselect_range(context = rs_get_context())
}
\arguments{
\item{first}{(integer) \cr
Either index of the first row to select or a vector of indices.
If \code{length(first) == 0} then current selection is deselected.
If \code{last} is not \code{NULL}, only the first value is used.}

\item{last}{(integer) \cr
Index of the last row to select or \code{NULL}. If not \code{NULL},
all lines from the \code{first[1]} to the \code{last[1]} are selected.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Select rows
}
